+++ SupportCompanion/Views/Applications.swift
@@ -43,6 +43,13 @@ struct Applications: View {
                     .ignoresSafeArea() // Ensure it covers the entire screen
                 } else {
                     ScrollView {
+						let mode = appState.preferences.mode
+						if mode != "SystemProfiler" {
+							Text("This list shows applications installed by \(mode).")
+								.font(.caption)
+								.foregroundColor(.secondary)
+								.padding(.bottom, 10)
+						}
                         LazyVGrid(
                             columns: columns,
                             alignment: .leading
@@ -52,12 +59,13 @@ struct Applications: View {
                                     .fixedSize(horizontal: false, vertical: false) // Allow vertical expansion
                             }
                         }
+						.padding(.bottom, 5)
                     }
                 }
             }
         }
         .padding(.horizontal, 20)
-        .padding(.bottom, 20)
+        //.padding(.bottom, 20)
         .onAppear {
             isLoading = true
             let appInfoManager = ApplicationsInfoManager(appState: appState) // Local instance