+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [2.4.0] - 2025-11-17
+## [2.4.0] - 2025-12-12
 ### Added
 - WiFi SSID information is now included in the device information
 - New option to hide tray menu icon. This allows for using the desktop information window without displaying the tray menu icon. Example configuration:
@@ -31,10 +31,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 <key>JamfLogPollHours</key>
 <integer>46</integer>
 ```
+- Logging will now be done in a log file located at `~/Library/Logs/SupportCompanion/SupportCompanion.log` in addition to os log. This allows for easier troubleshooting of issues with the app. The log file will be rotated when it reaches 5 MB in size. Debug logging can be enabled by setting the `LogLevel` to `Debug` in the configuration. Example configuration:
+```xml
+<key>DebugLogging</key>
+<true/>
+```
 ### Changed
 - MDM info now uses multiple MDM profile names to be able to correctly display the enrollment time.
 - The privileged helper tool has been migrated to `SMAppService` in place of `SMJobBless` since `SMJobBless` is deprecated in macOS 14 and later.
 - If `BrandName` is configured, it will now be displayed in the desktop information window as well as the header instead of "Device Information".
+- A new localized message will be displayed in the applications view stating that the apps are installed by the `mode`. This is to clarify the view only displays apps installed by the MDM and not all apps installed on the device.
+- User info will now use OpenDirectory to gather user information instead of `finger` command.
+- A delay has been added to `InfoHelp` when hovering over the info icon to prevent accidental triggering of the help popup.
+- App update names line limit has been increased to `2` lines to prevent truncation of long app names.
 ### Fixed
 - File watcher would not correctly detect changes on custom JSON cards if the file was replaced instead of modified. This has been fixed by using a different method to monitor file changes.
 - The pending updates badge on `Software Updates` was transparent in the main app.